0.3v Major update!
by adding a second screen for selecting a city, I made data transfer via UserDefaults so that the data is always saved and the last city that was selected remains, I added an icon if it's windy. Made a separate CocoaPods for storing the city of their data(width and longitude geographically), added a time zone for DateFormatter(). There was a problem with the city - St. Petersburg, until I figured out what the problem is, it does not show weather data for this city, in the next update I will try to fix it and mark up the data in the buttons